backend/pyproject.toml
[project]
name = "adk-chat-service"
version = "1.0.0"
description = "FastAPI backend for Google ADK chat integration with MCP extensibility"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.118.0",
    "uvicorn[standard]==0.34.0",
    "pydantic==2.11.1",
    "pydantic-settings==2.6.0",
    "google-adk==1.20.0",
    "google-generativeai==0.8.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "aiofiles==24.1.0",
    "httpx==0.28.1",
    "structlog==24.4.0",
    "websockets==13.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
